requirements.txt
discord
loguru
asyncio
scispacy
spacy
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.4.0/en_core_sci_sm-0.4.0.tar.gz
pyrebase4